APM-1744 Refactor to findCacheEntry
