+++ s3CacheTask/s3Utils/findCacheEntry.js
@@ -1,6 +1,14 @@
 const AWS = require('aws-sdk');
+const tar = require('tar-fs');
+const stream = require("stream");
+const { promisify } = require('util')
+const pipeline = promisify(stream.pipeline);
 
-const retrieveCacheEntry = async (keyName, bucketName, credentials) => {
+const extractCacheEntry = async (destination, cacheData) => {
+   await pipeline(cacheData.createReadStream(), tar.extract(destination));
+};
+
+const findCacheEntry = async (keyName, bucketName, credentials) => {
     try {
         const useLocal = process.env.AWS_ENV === 'localstack';
         const endpoint = useLocal ? 'http://localhost:4566' : undefined;
@@ -16,11 +24,11 @@ const retrieveCacheEntry = async (keyName, bucketName, credentials) => {
               Bucket: bucketName,
               Key: keyName
            }
-        )
+        );
         
      } catch (err) {
         return err;
      }
 };
 
-module.exports = retrieveCacheEntry;
+module.exports = { extractCacheEntry, findCacheEntry };
